// 3ME_UI/src/components/myPatients/allAlerts/missingRx/missingRx.enum.ts
export enum MissingRxSection {
  E_PRESBRIBER = "eprescriber",
  E_PRESBRIBER_SENT = "eprescribersent",
  RX_UPLOAD = "rxuploadprescriber",
  RX_UPLOAD_SENT = "rxuploadprescribersent",
  RX_UPLOAD_FAILED = "rxuploadprescriberfailed",
  FAX_IN_LATER = "fax-in-later",
  FAX_IN_LATER_SENT = "fax-in-later-sent",
  CHANGE_PRESCRIBER_TYPE = "change-prescriber",
  MISSING_RX_LOADER = "missingrxloading",
}